change 2 16bit gray img

+++ src/super_gradients/training/transforms/transforms.py
@@ -529,7 +529,7 @@ class DetectionMosaic(AbstractDetectionTransform, LegacyDetectionTransformMixin)
                 # generate output mosaic image
                 (h, w, c) = img.shape[:3]
                 if i_mosaic == 0:
-                    mosaic_img = np.full((input_h * 2, input_w * 2, c), self.border_value, dtype=np.uint8)
+                    mosaic_img = np.full((input_h * 2, input_w * 2, c), self.border_value, dtype=np.float32)
 
                 # suffix l means large image, while s means small image in mosaic aug.
                 (l_x1, l_y1, l_x2, l_y2), (s_x1, s_y1, s_x2, s_y2) = get_mosaic_coordinate(i_mosaic, xc, yc, w, h, input_h, input_w)
